# docs/openapi/digigraph.json
{
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "additionalProperties": false,
        "description": "OpenAI POST /v1/chat/completions request.",
        "properties": {
          "allowed_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional tool allowlist for this completion. Overrides project/env when set. Also accepted via X-Allowed-Tools header (comma-separated).",
            "title": "Allowed Tools"
          },
          "enable_web_search": {
            "default": false,
            "description": "Opt-in public web search via the digigraph ``web_search`` tool (digillm). Default off \u2014 corpus-only. Also accepted via X-Digi-Enable-Web-Search. When off, the model must not call web; when on, External cites supplement vault/search hits and never replace them (#3420).",
            "title": "Enable Web Search",
            "type": "boolean"
          },
          "force_tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional per-request locate tool to run with the user string as its query. Also accepted via X-Digi-Force-Tool. Aliases: search/digisearch, docs/digivault. Injected \u2014 the model is not asked to write the query.",
            "title": "Force Tool"
          },
          "messages": {
            "description": "Conversation messages",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "default": "digigraph-rag",
            "description": "Model id (ignored; we use project config)",
            "title": "Model",
            "type": "string"
          },
          "openwebui_format": {
            "default": false,
            "description": "If true, format tool blocks for Open WebUI (<details>, summary + tables). Also enabled by X-Response-Format: openwebui. model=digigraph-rag alone does not enable this; opt out anytime via X-Suppress-Tool-Stream or X-Response-Format: plain|neutral|none|digichat.",
            "title": "Openwebui Format",
            "type": "boolean"
          },
          "require_tool_calls": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional per-request signal that this completion needs tool_choice='required'. Also accepted via X-Require-Tool-Calls header. Combined with project agents.require_tool_calls and env DIGI_REQUIRE_TOOL_CALLS as a FLOOR (any true value wins) \u2014 unlike allowed_tools, this can only raise the requirement, never lower one the deployment already mandates.",
            "title": "Require Tool Calls"
          },
          "research_system_prompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Deprecated and IGNORED. Accepted only so clients that still send it do not get a 422; its value never reaches graph state. The research system prompt is operator-configured only \u2014 project config (agents.research_system_prompt) or the tenant corpus map (DIGI_TENANT_CORPUS_MAP) \u2014 so a caller cannot inject its own system prompt (CWE-639 / prompt injection).",
            "title": "Research System Prompt"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional conversation/session id. Isolates digistore and checkpoint state per conversation. Also set via X-Session-Id or X-Thread-Id header.",
            "title": "Session Id"
          },
          "stream": {
            "default": false,
            "description": "If true, return SSE stream",
            "title": "Stream",
            "type": "boolean"
          }
        },
        "required": [
          "messages"
        ],
        "title": "ChatCompletionRequest",
        "type": "object"
      },
      "ChatMessage": {
        "description": "OpenAI-style message.",
        "properties": {
          "content": {
            "default": "",
            "description": "Message content (string or OpenAI/AI SDK part list)",
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "user, assistant, or system",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowResult": {
        "description": "Result of run_digigraph_workflow. Phase 0: backtest result only.",
        "properties": {
          "backtest_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "digiquant BacktestResult when workflow ran a backtest",
            "title": "Backtest Result"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stable machine code for digichat (e.g. free_quota_exceeded); None on success",
            "title": "Error Code"
          },
          "message": {
            "default": "",
            "description": "Human-readable summary",
            "title": "Message",
            "type": "string"
          },
          "optimize_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error from optimize step without failing whole workflow",
            "title": "Optimize Error"
          },
          "optimize_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "digiquant OptimizeResult when optimize step ran",
            "title": "Optimize Result"
          },
          "profiling_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Merged profiling questions (brief + trading profile gaps)",
            "title": "Profiling Questions"
          },
          "rag_sources": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregated digisearch citations from the research step",
            "title": "Rag Sources"
          },
          "research_brief": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured research brief when research subgraph produced one",
            "title": "Research Brief"
          },
          "success": {
            "description": "Whether the workflow completed successfully",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "WorkflowResult",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Orchestration brain for digithings: LangGraph workflows, OpenAI-compatible chat, and federated vertical tools (digisearch, digiquant, digivault). Interactive docs: `/docs` (Swagger) and `/redoc`.",
    "title": "digigraph",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/threads/{thread_id}/state": {
      "get": {
        "description": "Return current (or specified) checkpoint state for a thread.\nRequires a checkpointer (default: memory when DIGI_CHECKPOINTER unset). Returns stored_datasets, research_response, error, etc.",
        "operationId": "get_thread_state_threads__thread_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "checkpoint_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Checkpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread State"
      }
    },
    "/v1/chat/completions": {
      "post": {
        "description": "OpenAI-compatible chat completions. Runs RAG workflow (LLM + search) and returns\nthe response as a chat message. Use as a model in Open WebUI.\nWhen stream=true: progressive SSE with tool-call blocks then final answer.\nTo get Open WebUI\u2013style tool blocks (<details>, markdown tables), send header\nX-Response-Format: openwebui or body openwebui_format=true. model=digigraph-rag alone\ndoes not enable that chrome. digichat sends X-Suppress-Tool-Stream /\nX-Response-Format: plain as belt-and-suspenders; activity arrives via digigraph_trace.",
        "operationId": "chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Completions",
        "tags": [
          "openai-compatible"
        ]
      }
    },
...
    "/workflow": {
      "post": {
        "description": "digiclaw custom skill: run_digigraph_workflow.\nPhase 0: user idea \u2192 backtest via digiquant \u2192 result in < 10s.",
        "operationId": "run_digigraph_workflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run digigraph workflow"
      }
    }
  }
}